requirements.txt
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
certifi==2017.11.5
chardet==3.0.4
decorator==4.1.2
Django==1.11.7
enum34==1.1.6
idna==2.6
influxdb==4.1.1
ipdb==0.10.3
ipython==5.5.0
ipython-genutils==0.2.0
oauthlib==2.0.6
pathlib2==2.3.0
pexpect==4.3.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
python-dateutil==2.6.1
pytz==2017.3
requests==2.18.4
requests-oauthlib==0.8.0
scandir==1.6
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
tweepy==3.5.0
urllib3==1.22
wcwidth==0.1.7
websocket-client==0.44.0
